remove MonoInstallTrigger and UnityMonoPathProvider
